change to float the precision of downloaded data

+++ R/cdtDownloadRFE_chirps.R
@@ -375,7 +375,7 @@ chirps.extract.data <- function(dest, ncfl, bbox){
         dy <- ncdf4::ncdim_def("lat", "degreeN", y, longname = "Latitude")
         missval <- -99
         ncgrd <- ncdf4::ncvar_def("rfe", "mm", list(dx, dy), missval,
-                                  "Rainfall Estimate", "short",
+                                  "Rainfall Estimate", "float",
                                   compression = 9)
         z[is.na(z)] <- missval
